split enum classes, fixed exchanger test case
