--- kahypar/partition/context/coarsening_enum_classes.h
@@ -1,7 +1,0 @@
-enum class CoarseningAlgorithm : uint8_t {
-  heavy_full,
-  heavy_lazy,
-  ml_style,
-  do_nothing,
-  UNDEFINED
-};
