+++ kahypar/partition/context/enum_classes/coarsening_enum_classes.h
@@ -1,0 +1,39 @@
+#pragma once
+#include <iostream>
+#include "kahypar/macros.h"
+namespace kahypar {
+
+enum class CoarseningAlgorithm : uint8_t {
+  heavy_full,
+  heavy_lazy,
+  ml_style,
+  do_nothing,
+  UNDEFINED
+};
+
+static CoarseningAlgorithm coarseningAlgorithmFromString(const std::string& type) {
+  if (type == "heavy_full") {
+    return CoarseningAlgorithm::heavy_full;
+  } else if (type == "heavy_lazy") {
+    return CoarseningAlgorithm::heavy_lazy;
+  } else if (type == "ml_style") {
+    return CoarseningAlgorithm::ml_style;
+  } else if (type == "do_nothing") {
+    return CoarseningAlgorithm::do_nothing;
+  }
+  LOG << "Illegal option:" << type;
+  exit(0);
+  return CoarseningAlgorithm::heavy_lazy;
+}
+std::ostream& operator<< (std::ostream& os, const CoarseningAlgorithm& algo) {
+  switch (algo) {
+    case CoarseningAlgorithm::heavy_full: return os << "heavy_full";
+    case CoarseningAlgorithm::heavy_lazy: return os << "heavy_lazy";
+    case CoarseningAlgorithm::ml_style: return os << "ml_style";
+    case CoarseningAlgorithm::do_nothing: return os << "do_nothing";
+    case CoarseningAlgorithm::UNDEFINED: return os << "UNDEFINED";
+      // omit default case to trigger compiler warning for missing cases
+  }
+  return os << static_cast<uint8_t>(algo);
+}
+} // namespace kahypar
