+++ kahypar/partition/context/enum_classes/initial_partitioning_enum_classes.h
@@ -1,0 +1,102 @@
+#pragma once
+#include <iostream>
+#include "kahypar/macros.h"
+namespace kahypar {
+enum class InitialPartitionerAlgorithm : uint8_t {
+  greedy_sequential,
+  greedy_global,
+  greedy_round,
+  greedy_sequential_maxpin,
+  greedy_global_maxpin,
+  greedy_round_maxpin,
+  greedy_sequential_maxnet,
+  greedy_global_maxnet,
+  greedy_round_maxnet,
+  bfs,
+  random,
+  lp,
+  pool,
+  UNDEFINED
+};
+
+enum class InitialPartitioningTechnique : uint8_t {
+  multilevel,
+  flat,
+  UNDEFINED
+};
+
+static InitialPartitionerAlgorithm initialPartitioningAlgorithmFromString(const std::string& mode) {
+  if (mode == "greedy_sequential") {
+    return InitialPartitionerAlgorithm::greedy_sequential;
+  } else if (mode == "greedy_global") {
+    return InitialPartitionerAlgorithm::greedy_global;
+  } else if (mode == "greedy_round") {
+    return InitialPartitionerAlgorithm::greedy_round;
+  } else if (mode == "greedy_sequential_maxpin") {
+    return InitialPartitionerAlgorithm::greedy_sequential_maxpin;
+  } else if (mode == "greedy_global_maxpin") {
+    return InitialPartitionerAlgorithm::greedy_global_maxpin;
+  } else if (mode == "greedy_round_maxpin") {
+    return InitialPartitionerAlgorithm::greedy_round_maxpin;
+  } else if (mode == "greedy_sequential_maxnet") {
+    return InitialPartitionerAlgorithm::greedy_sequential_maxnet;
+  } else if (mode == "greedy_global_maxnet") {
+    return InitialPartitionerAlgorithm::greedy_global_maxnet;
+  } else if (mode == "greedy_round_maxnet") {
+    return InitialPartitionerAlgorithm::greedy_round_maxnet;
+  } else if (mode == "lp") {
+    return InitialPartitionerAlgorithm::lp;
+  } else if (mode == "bfs") {
+    return InitialPartitionerAlgorithm::bfs;
+  } else if (mode == "random") {
+    return InitialPartitionerAlgorithm::random;
+  } else if (mode == "pool") {
+    return InitialPartitionerAlgorithm::pool;
+  }
+  LOG << "Illegal option:" << mode;
+  exit(0);
+  return InitialPartitionerAlgorithm::greedy_global;
+}
+
+static InitialPartitioningTechnique inititalPartitioningTechniqueFromString(const std::string& technique) {
+  if (technique == "flat") {
+    return InitialPartitioningTechnique::flat;
+  } else if (technique == "multi") {
+    return InitialPartitioningTechnique::multilevel;
+  }
+  LOG << "Illegal option:" << technique;
+  exit(0);
+  return InitialPartitioningTechnique::multilevel;
+}
+
+std::ostream& operator<< (std::ostream& os, const InitialPartitionerAlgorithm& algo) {
+  switch (algo) {
+    case InitialPartitionerAlgorithm::greedy_sequential: return os << "greedy_sequential";
+    case InitialPartitionerAlgorithm::greedy_global: return os << "greedy_global";
+    case InitialPartitionerAlgorithm::greedy_round: return os << "greedy_round";
+    case InitialPartitionerAlgorithm::greedy_sequential_maxpin: return os << "greedy_maxpin";
+    case InitialPartitionerAlgorithm::greedy_global_maxpin: return os << "greedy_global_maxpin";
+    case InitialPartitionerAlgorithm::greedy_round_maxpin: return os << "greedy_round_maxpin";
+    case InitialPartitionerAlgorithm::greedy_sequential_maxnet: return os << "greedy_maxnet";
+    case InitialPartitionerAlgorithm::greedy_global_maxnet: return os << "greedy_global_maxnet";
+    case InitialPartitionerAlgorithm::greedy_round_maxnet: return os << "greedy_round_maxnet";
+    case InitialPartitionerAlgorithm::bfs: return os << "bfs";
+    case InitialPartitionerAlgorithm::random: return os << "random";
+    case InitialPartitionerAlgorithm::lp: return os << "lp";
+    case InitialPartitionerAlgorithm::pool: return os << "pool";
+    case InitialPartitionerAlgorithm::UNDEFINED: return os << "UNDEFINED";
+      // omit default case to trigger compiler warning for missing cases
+  }
+  return os << static_cast<uint8_t>(algo);
+}
+
+std::ostream& operator<< (std::ostream& os, const InitialPartitioningTechnique& technique) {
+  switch (technique) {
+    case InitialPartitioningTechnique::flat: return os << "flat";
+    case InitialPartitioningTechnique::multilevel: return os << "multilevel";
+    case InitialPartitioningTechnique::UNDEFINED: return os << "UNDEFINED";
+      // omit default case to trigger compiler warning for missing cases
+  }
+  return os << static_cast<uint8_t>(technique);
+}
+} // namespace kahypar
