+++ kahypar/partition/context/include_parameters.h
@@ -1,8 +1,9 @@
 #pragma once
-#include "kahypar/partition/context/context_enum_classes.h"
+#include "kahypar/partition/context/enum_classes/context_enum_classes.h"
 #include "kahypar/partition/context/coarsening_parameters.h"
 #include "kahypar/partition/context/evolutionary_parameters.h"
 #include "kahypar/partition/context/initial_partitioning_parameters.h"
 #include "kahypar/partition/context/local_search_parameters.h"
 #include "kahypar/partition/context/partitioning_parameters.h"
+
 #include "kahypar/partition/context/rating_parameters.h"
