+++ kahypar/partition/context/local_search_parameters.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "kahypar/partition/context/context_enum_classes.h"
+#include "kahypar/partition/context/enum_classes/local_search_enum_classes.h"
 namespace kahypar {
 struct LocalSearchParameters {
   struct FM {
