+++ tests/partition/evolutionary/probability_table_test.cc
@@ -21,7 +21,7 @@
 #include "gmock/gmock.h"
 
 #include "kahypar/partition/context.h"
-#include "kahypar/partition/context_enum_classes.h"
+#include "kahypar/partition/context/enum_classes/evolutionary_enum_classes.h"
 #include "kahypar/partition/evolutionary/probability_tables.h"
 #include "kahypar/utils/randomize.h"
